+++ src/main/java/tap/execounting/components/grids/LessonGrid.java
@@ -2,7 +2,6 @@ package tap.execounting.components.grids;
 
 import java.util.List;
 
-
 import org.apache.tapestry5.ComponentResources;
 import org.apache.tapestry5.annotations.InjectComponent;
 import org.apache.tapestry5.annotations.Property;
@@ -20,6 +19,7 @@ import tap.execounting.entities.EventType;
 import tap.execounting.entities.Facility;
 import tap.execounting.entities.Room;
 import tap.execounting.entities.Teacher;
+import tap.execounting.security.AuthorizationDispatcher;
 
 public class LessonGrid {
 	@Inject
@@ -28,9 +28,12 @@ public class LessonGrid {
 	private ComponentResources componentResources;
 	@Inject
 	private CrudServiceDAO dao;
+	@Inject
+	@Property
+	private AuthorizationDispatcher dispatcher;
 	@InjectComponent
 	private Zone ezone;
-	
+
 	@InjectComponent
 	private AddEvent editor;
 	@Property
@@ -40,24 +43,32 @@ public class LessonGrid {
 	@Property
 	private Event unit;
 
-	public List<Event> getSource(){
+	public List<Event> getSource() {
 		return dao.findWithNamedQuery(Event.ALL);
 	}
-	
+
 	Object onActionFromEdit(Event c) {
-		editorActive = true;
-		editor.setup(c, true);
+		// AUTHORIZATION MOUNT POINT
+		if (dispatcher.canEditEvents()) {
+			editorActive = true;
+			editor.setup(c, true);
+		}
 		return ezone.getBody();
 	}
 
 	Object onActionFromAdd() {
-		editorActive = true;
-		editor.reset();
+		// AUTHORIZATION MOUNT POINT
+		if (dispatcher.canCreateEvents()) {
+			editorActive = true;
+			editor.reset();
+		}
 		return ezone.getBody();
 	}
 
 	void onDelete(Event c) {
-		dao.delete(Event.class, c.getId());
+		// AUTHORIZATION MOUNT POINT
+		if (dispatcher.canDeleteEvents())
+			dao.delete(Event.class, c.getId());
 	}
 
 	void setupRender() {
@@ -69,42 +80,44 @@ public class LessonGrid {
 			model.exclude("id");
 		}
 	}
-	
-	public String getHostName(){
+
+	public String getHostName() {
 		System.out.println(unit.getId());
 		return dao.find(Teacher.class, unit.getHostId()).getName();
 	}
-	
-	public String getFacilityTitle(){
+
+	public String getFacilityTitle() {
 		return dao.find(Facility.class, unit.getFacilityId()).getName();
 	}
-	
-	public String getRoomTitle(){
+
+	public String getRoomTitle() {
 		return dao.find(Room.class, unit.getRoomId()).getName();
 	}
-	
-	public String getState(){
-		return unit.getState() == EventState.complete ? "Проведено" : "Не проведено";
+
+	public String getState() {
+		return unit.getState() == EventState.complete ? "Проведено"
+				: "Не проведено";
 	}
-	
-	public String getTypeTitle(){
+
+	public String getTypeTitle() {
 		return dao.find(EventType.class, unit.getTypeId()).getTitle();
 	}
-	
+
 	public int getMoney() {
 		EventType et = dao.find(EventType.class, unit.getTypeId());
 		if (unit.getClients().size() > 0)
 			return et.getMoney() * unit.getClients().size();
 		return et.getMoney();
 	}
-	
-	public String getClients(){
+
+	public String getClients() {
 		StringBuilder sb = new StringBuilder();
 		List<Client> lc = unit.getClients();
-		
-		for(int i = 0;i<lc.size();i++){
+
+		for (int i = 0; i < lc.size(); i++) {
 			sb.append(lc.get(i).getName());
-			if(i<lc.size()-1) sb.append(", ");
+			if (i < lc.size() - 1)
+				sb.append(", ");
 		}
 		return sb.toString();
 	}
